Refactor: 애플 v1.1 컨트롤러 작성 및 리팩토링
