+++ src/main/java/com/example/d3nserver/auth/apple/controller/AppleController.java
@@ -1,7 +1,8 @@
-package com.example.d3nserver.auth.apple;
+package com.example.d3nserver.auth.apple.controller;
 
 import com.example.d3nserver.auth.apple.request.AppleLoginRequest;
-import com.example.d3nserver.auth.common.response.AuthResponse;
+import com.example.d3nserver.auth.apple.service.AppleAuthService;
+import com.example.d3nserver.auth.authToken.response.AuthResponse;
 import com.example.d3nserver.auth.jwt.ReqUser;
 import com.example.d3nserver.common.base.BaseResponse;
 import com.example.d3nserver.user.domain.User;
