+++ src/main/java/com/example/d3nserver/auth/apple/service/AppleAuthService.java
@@ -1,11 +1,10 @@
-package com.example.d3nserver.auth.apple;
+package com.example.d3nserver.auth.apple.service;
 
 import com.example.d3nserver.auth.apple.request.AppleLoginRequest;
 import com.example.d3nserver.auth.apple.response.AppleAuthTokenResponse;
-import com.example.d3nserver.auth.common.response.AuthResponse;
+import com.example.d3nserver.auth.authToken.response.AuthResponse;
 import com.example.d3nserver.auth.jwt.authToken.AuthToken;
 import com.example.d3nserver.auth.jwt.provider.AuthTokenProvider;
-import com.example.d3nserver.common.base.BaseException;
 import com.example.d3nserver.common.util.parser.JwtParser;
 import com.example.d3nserver.user.domain.MemberProvider;
 import com.example.d3nserver.user.domain.RoleType;
@@ -13,7 +12,6 @@ import com.example.d3nserver.user.domain.User;
 import com.example.d3nserver.user.service.UserService;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
 import org.bouncycastle.openssl.PEMParser;
@@ -33,9 +31,7 @@ import org.springframework.web.client.RestTemplate;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
-import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
-import java.security.spec.InvalidKeySpecException;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.*;
